[#20] revert 🔧 vite-plugin-dts 로 통일
