+++ index.ts
@@ -1,0 +1,10 @@
+// only for vite, tsup
+// remember, this is not barrel file.
+const moduleMap = {
+    isEmpty: './src/isEmpty.ts',
+    size: './src/size.ts',
+    keys: './src/keys.ts',
+    isNil: './src/isNil.ts',
+} as const
+
+export default moduleMap
