+++ package.json
@@ -89,7 +89,6 @@
         "rollup": "^4.24.3",
         "rollup-preserve-directives": "^1.1.2",
         "vite": "^5.4.10",
-        "vite-plugin-dts": "^4.3.0",
         "vite-tsconfig-paths": "^5.0.1",
         "vitest": "^2.1.4"
     },
